requirements.txt
Django==3.2.8
psycopg2-binary==2.8.6
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4